# api/swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Cities App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /:
    x-swagger-router-controller: cities
    get:
      tags:
        - "city"
      description: Returns 'Cities'
      # used as the method name of the controller
      operationId: Cities
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CitiesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - "city"
      description: "create new City"
      operationId: "createCity"
      consumes:
        - "application/json"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CitiesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  CitiesResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string